+++ src/Command/DeliveryStatusChangeCommand.php
@@ -39,6 +39,7 @@ class DeliveryStatusChangeCommand extends Command
             ->addArgument('identifier', InputArgument::REQUIRED, 'Identifier of the request recipient or status change')
             ->addOption('status-type', null, InputOption::VALUE_OPTIONAL, 'The status type to set')
             ->addOption('description', null, InputOption::VALUE_OPTIONAL, 'The description to set')
+            ->addOption('set-end-date', null, InputOption::VALUE_NONE, 'Set delivery_end_date to current date if not already set and the created status is final')
             ->addOption('with-file', null, InputOption::VALUE_NONE, 'Wether to attach a file to the status change');
     }
 
@@ -47,6 +48,7 @@ class DeliveryStatusChangeCommand extends Command
         $action = $input->getArgument('action');
         $identifier = $input->getArgument('identifier');
         $withFile = (bool) $input->getOption('with-file');
+        $setEndDate = (bool) $input->getOption('set-end-date');
         $statusType = (int) ($input->getOption('status-type') ?? Vendo::getStatusForCode('P6'));
         $description = $input->getOption('description') ?? 'Test status change';
 
@@ -62,9 +64,9 @@ class DeliveryStatusChangeCommand extends Command
                 $output->writeln('Generating DeliveryStatusChange for request recipient "'.$identifier.'"...');
                 $deliveryStatusChange = $this->dispatchService->createDeliveryStatusChange($identifier, $statusType, $description, $fileData);
 
+                $recipient = $deliveryStatusChange->getDispatchRequestRecipient();
                 // Set delivery end date if status is final.
-                if ($deliveryStatusChange->isFinalDualDeliveryStatus()) {
-                    $recipient = $deliveryStatusChange->getDispatchRequestRecipient();
+                if ($setEndDate && !$recipient->getDeliveryEndDate() && $deliveryStatusChange->isFinalDualDeliveryStatus()) {
                     $recipient->setDeliveryEndDate(new \DateTimeImmutable('now', new \DateTimeZone('UTC')));
                     $this->dispatchService->updateRequestRecipient($recipient);
                 }